Showing users' own photos

+++ app/controllers/likes_controller.rb
@@ -24,9 +24,9 @@ class LikesController < ApplicationController
 
     if the_like.valid?
       the_like.save
-      redirect_to("/likes", { :notice => "Like created successfully." })
+      redirect_to("/photos/#{the_like.photo_id}", { :notice => "Like created successfully." })
     else
-      redirect_to("/likes", { :alert => the_like.errors.full_messages.to_sentence })
+      redirect_to("/photos/#{the_like.photo_id}", { :alert => the_like.errors.full_messages.to_sentence })
     end
   end
 
@@ -49,8 +49,14 @@ class LikesController < ApplicationController
     the_id = params.fetch("path_id")
     the_like = Like.where({ :id => the_id }).at(0)
 
-    the_like.destroy
+    the_like.fan_id = params.fetch("query_fan_id")
+    the_like.photo_id = params.fetch("query_photo_id")
 
-    redirect_to("/likes", { :notice => "Like deleted successfully."} )
+    if the_like.valid?
+      # redirect_to("/photos/#{the_id}", { :notice => "Like deleted successfully."} )
+      the_like.destroy
+    else
+      redirect_to("/likes/#{the_like.id}", { :alert => the_like.errors.full_messages.to_sentence })
+    end
   end
 end
